+++ R/prv_build_dictionary.R
@@ -22,6 +22,7 @@ build_dictionary <- function(DF, dictlist) {
 
     # Extra details for the data dictionary
     dct_type_strs <- sapply(DF, coltype)
+    dct_ordered   <- sapply(DF, is.ordered)
     dct_classes   <- sapply(DF, class)
     dct_types     <- sapply(DF, typeof)
     dct_pct_miss  <- sapply(DF, function(col) { trunc((sum(is.na(col)) / length(col)) * 100) })
@@ -47,6 +48,7 @@ build_dictionary <- function(DF, dictlist) {
             all_same    = dct_all_same,
             val_lab     = codify(raw_val_labs),
             fct_lvl     = codify(raw_fct_lvls),
+            fct_ordered = dct_ordered,
             class       = codify(dct_classes),
             type        = codify(dct_types),
             haystack    = haystacks