+++ NEWS.md
@@ -4,6 +4,7 @@
 # sift 0.2.0 (2023-03-27)
 
 - ADDED
+    - `save_dictionary()` allows you to save a data dictionary in a form that my other package, [`tsv2label`](https://github.com/DesiQuintans/tsv2label), will accept. Closes #11.
     - `options_sift()` prints the status of all options when invoked with no arguments. Closes #12.
 
 - FIXED